sync revision 1553 	- buffer address, map & unmap 	- build shaders with Visual Studio 	- initial material system implementation 	- handle window resize 	- support secondary command buffers with the Vulkan backend 	- replaced cpu count with core count & thread count 	- CMake Windows build support 	- replaced placeholder scene 	- per-frame memory heap 	- updated PhysFS with the latest upstream changes 	- updated Lua to 5.4.3

+++ Drivers/Vulkan/Shaders/Texture.h
@@ -8,16 +8,15 @@ layout(set = 0, binding = 0) uniform sampler Re_samplers[];
 layout(set = 0, binding = 1) uniform texture2D Re_textures[];
 
 vec4
-Re_SampleSceneTexture(uint16_t id, vec2 uv)
+Re_SampleSceneTexture(uint id, vec2 uv)
 {
-	return texture(sampler2D(Re_textures[uint(id)], Re_samplers[SCENE_SAMPLER]), uv);
+	return texture(sampler2D(Re_textures[id], Re_samplers[SCENE_SAMPLER]), uv);
 }
 
 vec4
-Re_SampleUITexture(uint16_t id, vec2 uv)
+Re_SampleUITexture(uint id, vec2 uv)
 {
-	return texture(sampler2D(Re_textures[uint(id)], Re_samplers[UI_SAMPLER]), uv);
+	return texture(sampler2D(Re_textures[id], Re_samplers[UI_SAMPLER]), uv);
 }
 
 #endif /* _RE_TEXTURES_H_ */
-
